pyproject.toml
[project]
name = "radius-portal"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "Flask>=3.0",
  "psycopg[binary]>=3.1",
  "psycopg[pool]>=3.3.0",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.4",
]

[tool.ruff]
line-length = 100

